requirements.txt
# =============================================================================
# Core Framework Dependencies
# =============================================================================
fastapi==0.115.11
uvicorn==0.34.0
starlette==0.46.0
h11==0.14.0
anyio==4.8.0
sniffio==1.3.1

# =============================================================================
# Data Processing & Parsing
# =============================================================================
beautifulsoup4==4.13.3
feedparser==6.0.11
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
sgmllib3k==1.0.0
soupsieve==2.6

# =============================================================================
# AI/ML Dependencies
# =============================================================================
openai==1.64.0
tenacity==9.0.0

# =============================================================================
# Development & Testing Tools
# =============================================================================
black==25.1.0
coverage==7.6.12
pytest==8.3.4
pytest-cov==6.0.0
pylint==3.3.4
astroid==3.3.8
dill==0.3.9
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
tomlkit==0.13.2

# =============================================================================
# Utility Libraries
# =============================================================================
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
python-dotenv==1.0.1
schedule==1.2.2
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0

# =============================================================================
# Additional Dependencies
# =============================================================================
psutil==5.9.5
